pyproject.toml
[project]
name = "auralis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdullah-1121", email = "muhammad11abdullah21@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "weave>=0.51.56",
]

[project.scripts]
auralis = "auralis:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
